feat: refactor cutler to have a print-versions command
I also wrote up what I think our repo branching strategy could
look like.
